requirements.txt
azure-functions
pandas
openpyxl
azure-ai-projects
azure-identity
azure-storage-blob
python-dotenv